代码优化

+++ dorm-service/src/main/java/com/waken/dorm/serviceImpl/dorm/DormBuildingServiceImpl.java
@@ -2,13 +2,14 @@ package com.waken.dorm.serviceImpl.dorm;
 
 import com.baomidou.mybatisplus.core.conditions.query.LambdaQueryWrapper;
 import com.baomidou.mybatisplus.core.metadata.IPage;
+import com.baomidou.mybatisplus.extension.service.impl.ServiceImpl;
 import com.waken.dorm.common.constant.Constant;
-import com.waken.dorm.common.entity.dorm.DormBuilding;
+import com.waken.dorm.common.entity.dorm.Building;
 import com.waken.dorm.common.enums.CodeEnum;
 import com.waken.dorm.common.exception.ServerException;
 import com.waken.dorm.common.form.base.DeleteForm;
-import com.waken.dorm.common.form.dorm.DormBuildingForm;
-import com.waken.dorm.common.form.dorm.EditDormBuildingForm;
+import com.waken.dorm.common.form.dorm.BuildingForm;
+import com.waken.dorm.common.form.dorm.EditBuildingForm;
 import com.waken.dorm.common.manager.UserManager;
 import com.waken.dorm.common.utils.Assert;
 import com.waken.dorm.common.utils.BeanMapper;
@@ -25,6 +26,7 @@ import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Propagation;
 import org.springframework.transaction.annotation.Transactional;
 
+import java.util.ArrayList;
 import java.util.List;
 
 /**
@@ -37,7 +39,7 @@ import java.util.List;
 @Service
 @RequiredArgsConstructor(onConstructor = @__(@Autowired))
 @Transactional(propagation = Propagation.SUPPORTS, readOnly = true, rollbackFor = Exception.class)
-public class DormBuildingServiceImpl implements DormBuildingService {
+public class DormBuildingServiceImpl extends ServiceImpl<DormBuildingMapper, Building> implements DormBuildingService {
     private final DormBuildingMapper buildingMapper;
 
     /**
@@ -48,23 +50,23 @@ public class DormBuildingServiceImpl implements DormBuildingService {
      */
     @Transactional
     @Override
-    public DormBuilding saveDormBuilding(EditDormBuildingForm editBuildingForm) {
+    public Building saveDormBuilding(EditBuildingForm editBuildingForm) {
         this.editBuildingValidate(editBuildingForm);
         int count;
-        DormBuilding dormBuilding = new DormBuilding();
-        BeanMapper.copy(editBuildingForm, dormBuilding);
-        if (StringUtils.isEmpty(editBuildingForm.getPkDormBuildingId())) {//新增
+        Building building = new Building();
+        BeanMapper.copy(editBuildingForm, building);
+        if (StringUtils.isEmpty(editBuildingForm.getId())) {//新增
             log.info("service: 开始进入新增宿舍楼信息");
-            dormBuilding.setStatus(CodeEnum.ENABLE.getCode());
-            count = buildingMapper.insert(dormBuilding);
+            building.setIsDeleted(false);
+            count = buildingMapper.insert(building);
             if (count == Constant.ZERO) {
                 throw new ServerException("新增个数为 0 条");
             }
-            return dormBuilding;
+            return building;
         } else {//更新宿舍楼信息
             log.info("service: 开始进入更新宿舍楼信息");
-            buildingMapper.updateById(dormBuilding);
-            return buildingMapper.selectById(editBuildingForm.getPkDormBuildingId());
+            buildingMapper.updateById(building);
+            return buildingMapper.selectById(editBuildingForm.getId());
         }
     }
 
@@ -78,21 +80,24 @@ public class DormBuildingServiceImpl implements DormBuildingService {
     public void deleteDormBuilding(DeleteForm deleteForm) {
         log.info("service: 删除宿舍楼开始");
         List<String> ids = deleteForm.getDelIds();
+        if (ids == null || ids.isEmpty()) {
+            return;
+        }
         Integer delStatus = deleteForm.getDelStatus();
         int count;
         if (CodeEnum.YES.getCode() == delStatus) { // 物理删除
-            List<DormBuilding> dormBuildingList = buildingMapper.selectByIds(ids);
-            StringBuffer sb = new StringBuffer();
-            for (DormBuilding dormBuilding : dormBuildingList) {
-                if (CodeEnum.ENABLE.getCode() == dormBuilding.getStatus()) {
-                    sb.append(dormBuilding.getDormBuildingNum());
-                }
-            }
-            Assert.isNull(sb.toString(),"操作失败,以下宿舍楼处于生效中：" + sb.toString());
             //删除宿舍
             count = buildingMapper.deleteBatchIds(ids);
             Assert.isFalse(count == Constant.ZERO);
         } else if (CodeEnum.NO.getCode() == delStatus) {//状态删除
+            List<Building> buildingList = new ArrayList<>(ids.size());
+            ids.stream().forEach(id -> {
+                Building building = new Building();
+                building.setId(id);
+                building.setIsDeleted(true);
+                buildingList.add(building);
+            });
+            this.updateBatchById(buildingList);
             count = buildingMapper.batchUpdateStatus(DormUtil.getToUpdateStatusMap(ids,UserManager.getCurrentUserId()));
             Assert.isFalse(count == Constant.ZERO);
         } else {
@@ -107,9 +112,9 @@ public class DormBuildingServiceImpl implements DormBuildingService {
      * @return
      */
     @Override
-    public IPage<DormBuildingView> listDormBuildings(DormBuildingForm buildingForm) {
+    public IPage<DormBuildingView> listDormBuildings(BuildingForm buildingForm) {
         log.info("service: 分页查询宿舍楼信息开始");
-        return buildingMapper.listDormBuildings(DataHandle.getWrapperPage(buildingForm),buildingForm);
+        return buildingMapper.findBuildingPage(DataHandle.getWrapperPage(buildingForm), buildingForm);
     }
 
     /**
@@ -117,23 +122,23 @@ public class DormBuildingServiceImpl implements DormBuildingService {
      *
      * @param editForm
      */
-    private void editBuildingValidate(EditDormBuildingForm editForm) {
-        if (StringUtils.isEmpty(editForm.getPkDormBuildingId())) {//新增验证
-            Assert.notNull(editForm.getDormBuildingType());
-            Assert.notNull(editForm.getDormBuildingNum());
-            Assert.notNull(editForm.getDormBuildingLevels());
-            DormBuilding dormBuilding = buildingMapper.selectOne(new LambdaQueryWrapper<DormBuilding>()
-                    .eq(DormBuilding::getDormBuildingNum, editForm.getDormBuildingNum())
+    private void editBuildingValidate(EditBuildingForm editForm) {
+        if (StringUtils.isEmpty(editForm.getId())) {//新增验证
+            Assert.notNull(editForm.getType());
+            Assert.notNull(editForm.getCode());
+            Assert.notNull(editForm.getFloorTotal());
+            Building building = buildingMapper.selectOne(new LambdaQueryWrapper<Building>()
+                    .eq(Building::getCode, editForm.getCode())
             );
-            Assert.isNull(dormBuilding,"已存在相同名称的楼栋号");
+            Assert.isNull(building, "已存在相同名称的楼栋号");
         } else {//修改验证
-            Assert.notNull(buildingMapper.selectById(editForm.getPkDormBuildingId()),"参数错误");
-            if (StringUtils.isNotEmpty(editForm.getDormBuildingNum())) {
-                DormBuilding dormBuilding = buildingMapper.selectOne(new LambdaQueryWrapper<DormBuilding>()
-                        .eq(DormBuilding::getDormBuildingNum, editForm.getDormBuildingNum())
+            Assert.notNull(buildingMapper.selectById(editForm.getId()), "参数错误");
+            if (StringUtils.isNotEmpty(editForm.getCode())) {
+                Building building = buildingMapper.selectOne(new LambdaQueryWrapper<Building>()
+                        .eq(Building::getCode, editForm.getCode())
                 );
-                if (null != dormBuilding) {
-                    if (!StringUtils.equals(dormBuilding.getPkDormBuildingId(), editForm.getPkDormBuildingId())) {
+                if (null != building) {
+                    if (!StringUtils.equals(building.getId(), editForm.getId())) {
                         throw new ServerException("已存在相同名称的楼栋号");
                     }
                 }
